\name{aftosmac-package}
\alias{aftosmac-package}
\alias{aftosmac}
\docType{package}
\title{
\packageTitle{aftosmac}
}
\description{
\packageDescription{aftosmac}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{aftosmac}
\packageIndices{aftosmac}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{aftosmac}

Maintainer: \packageMaintainer{aftosmac}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
