% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aftosmac.R
\name{aftosmac}
\alias{aftosmac}
\title{Optimal Subsampling Method for Accelerated Failure Time Models.}
\usage{
aftosmac(
  formula,
  data,
  r0,
  r,
  sspType,
  B = 1,
  R = 20,
  rankWt = c("gehan"),
  parDist = c("weibull"),
  eqType = c("s", "ns"),
  fitMtd = c("rank", "ls", "par"),
  estMtd = c("cs", "ce"),
  se = c("NULL", "parTrue", "parFull"),
  control = list()
)
}
\arguments{
\item{formula}{a formula expression, of the form \code{response ~ predictors}.
The \code{response} is a \code{Surv} object with right censoring.}

\item{data}{an optional data.frame in which to interpret the variables occurring
in the \code{formula}.}

\item{r0}{the pilot subsample size.}

\item{r}{the second-step subsample size.}

\item{sspType}{the type of optimal subsampling probabilities.}

\item{B}{number of subsmaples needed to derive the final estimator}

\item{R}{the number of noises used in the resampling approach.}

\item{rankWt}{the type of weight function when using rank-based semiparametric approach.}

\item{parDist}{the error distribution when using parametric approach.}

\item{eqType}{the type of estimating function when using rank-based semiparametric approach.}

\item{fitMtd}{the model to derive the estimator.}

\item{estMtd}{method to combine subsamples or combine estimators.}

\item{se}{method to calculate the standard errors.}

\item{control}{controls maxiter and tolerance.}
}
\value{
The function return a list containing at least the following components:
\describe{
  \item{coe}{a vector of point estimates}
  \item{std}{a vector of estimated standard errors}
  \item{converge}{indicator of convergence}
  \item{iter}{iteration needed to get the converging result}
}
}
\description{
Estimate the full data statistical inferences based on a subsample. The
subsample is derived by sampling with probabilities. The sampling probabilities
is defined by the A-optimality and L-optimality from optimal design of
experiments.
}
\keyword{aftosmac}
