% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aftosmac.R
\name{aftosmac}
\alias{aftosmac}
\title{Optimal Subsampling Method for Accelerated Failure Time Models.}
\usage{
aftosmac(
  formula,
  data,
  size.pilot,
  size.subsample,
  sspType = c("optA", "optL", "uniform"),
  model = c("weibull", "ls", "gehan"),
  se = c("NULL", "parTrue", "parFull"),
  combine = c("estimator", "sample"),
  repeated = 0,
  constrasts = NULL,
  control = list()
)
}
\arguments{
\item{formula}{a formula expression, of the form \code{response ~ predictors}.
The \code{response} is a \code{Surv} object with right censoring.}

\item{data}{an optional data.frame in which to interpret the variables occurring
in the \code{formula}.}

\item{size.pilot}{the pilot subsample size}

\item{size.subsample}{the second step subsample size}

\item{sspType}{the type of optimal subsampling probabilities.}

\item{model}{methods to fit the regression model.}

\item{se}{method to calculate the standard errors.}

\item{combine}{how to combine the pilot and second-step subsample.}

\item{repeated}{number of subsmaples needed to derive the final estimator}

\item{constrasts}{...}

\item{control}{controls maxiter and tolerance.}
}
\value{
The function return a list containing at least the following components:
\describe{
  \item{coe}{a vector of point estimates}
  \item{std}{a vector of estimated standard errors}
  \item{converge}{indicator of convergence}
  \item{iter}{iteration needed to get the converging result}
}
}
\description{
Estimate the full data statistical inferences based on a subsample. The
subsample is derived by sampling with probabilities. The sampling probabilities
is defined by the A-optimality and L-optimality from optimal design of
experiments.
}
\keyword{aftosmac}
